+++ src/rss_to_webhook/__main__.py
@@ -1,0 +1,6 @@
+"""Run the package as a script using `py -m rss_to_webhook`."""
+
+from rss_to_webhook.main import app
+
+if __name__ == "__main__":
+    app(prog_name="rick-portal-gun")