+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to the extension will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Package now executeable as a script. Try `rss-to-webhook --help` or `py -m rss_to_webhook --help`
+
 ## [0.0.3] - 2023-10-29
 
 ### Added